+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 + [x] Colors: improve work with hsl, hsv
 + [x] Spinner: fixed input not a number value, value more or less than defined in options.
 + [x] Keypad: fixed internal value type mismatch.
++ [x] InfoBox: added attribute `data-overlay-click-close=true|false`, default `false`, issue #1668
 
 ### 4.4.2
 #### Metro 4